[Close #222] New-post Action below PostsController
RSpec: 709 examples, 0 failures
RuboCop: 240 files inspected, no offenses detected
