--- lib/newpoc-action-post-new/.rspec
@@ -1,2 +1,0 @@
---format documentation
---color
