--- lib/newpoc-action-post-new/Gemfile
@@ -1,4 +1,0 @@
-source 'https://rubygems.org'
-
-# Specify your gem's dependencies in newpoc-action-post-new.gemspec
-gemspec
