--- lib/newpoc-action-post-new/lib/newpoc/action/post/new/version.rb
@@ -1,11 +1,0 @@
-
-module Newpoc
-  module Action
-    module Post
-      # New-post (pre-edit-attributes) encapsulation for `new_poc`.
-      class New
-        VERSION = '0.0.1'
-      end
-    end
-  end
-end
