--- lib/newpoc-action-post-new/spec/spec_helper.rb
@@ -1,9 +1,0 @@
-require 'simplecov'
-SimpleCov.start do # 'rails'
-  add_filter '/spec/'
-  add_filter '/vendor/'
-end
-
-$LOAD_PATH.unshift File.expand_path('../../lib', __FILE__)
-
-require 'newpoc/action/post/new'
